added args feed, name building, error handling, readme fix
