+++ valopers-query/README.md
@@ -1,11 +1,18 @@
 # Valopers query
-Creates four files with val. addresses. Each contain all, produced, claimed or waiting addresses.
+Creates four files with val. addresses. Each contains all, produced, claimed or waiting addresses.
 
 ### Setup
 `
+apt install python3
 pip3 install -r requirements.txt
 `
 ### Usage
 `
 python3 main.py
 `
+
+`
+python3 main.py path_to_files
+`
+The app accept url or absolute path to the file location.
+It also accept several paths (url, local) as sequnce of argq
